+++ CHANGELOG.md
@@ -17,8 +17,38 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
-## Unreleased
-## 0.1.11 (2024-10-14)
+## 0.1.12 (2025-02-19)
+---
+
+### New
+Implement Together AI provider and models [GS-158].
+Implement xAI Grok model [GS-157].
+Implement IBM watsonx provider [GS-155].
+Implement generic Langchain model abstract interface [GS-155].
+Implement Nvidia API / NIM / Nemotron [GS-93].
+Implement Rhymes.ai Aria chat model [GS-152].
+Implement Rhymes.ai Allegro video generation model (only configuration) [GS-153].
+Add AI_STREAMING envvar to configure the Streaming response method [GS-32].
+Add NVIDIA_API_KEY, NVIDIA_MODEL_NAME, NVIDIA_TEMPERATURE, NVIDIA_MAX_TOKENS, NVIDIA_TOP_P, and NVIDIA_BASE_URL envvars [GS-93].
+Add RHYMES_CHAT_API_KEY, RHYMES_CHAT_MODEL_NAME, RHYMES_CHAT_TEMPERATURE, RHYMES_CHAT_MAX_TOKENS, RHYMES_CHAT_TOP_P, RHYMES_CHAT_BASE_URL, RHYMES_VIDEO_API_KEY, RHYMES_VIDEO_MODEL_NAME, RHYMES_VIDEO_BASE_URL, RHYMES_VIDEO_NUM_STEP, and RHYMES_VIDEO_CFG_SCALE envvars [GS-152].
+Add AIMLAPI_TOP_P to configure top_p parameter in AI/ML API [GS-138].
+Add OPENAI_TOP_P to configure top_p parameter in OpenAI chat model.
+Add CustomLLM abstract class for LangChain to implement new LLM providers not implemented in LangChain yet [GS-155]. 
+
+### Changes
+Change "black-forest-labs/FLUX.1-schnell" image generation model by default.
+Change OPENAI_MAX_TOKENS and AIMLAPI_MAX_TOKENS to have '' by default to get the maximum tokens possible [GS-157].
+Change "grok-beta" changed to "grok-2" as default model for xAI [GS-157].
+Change "openai" instead of "openai_chat" to get the default OpenAI provider in the LANGCHAIN_DEFAULT_MODEL parameter.
+"get_openai_api()" function was added to standarize LLM client creation for providers compatible with the OpenAI completions API [GS-157].
+
+### Fixes
+Fix the "ValueError: invalid literal for int() with base 10: ''" error in get_vision_response() when OPENAI_MAX_TOKENS is empty [GS-152].
+Fix poetry 2.x "The option --no-update does not exist" error message [FA-84].
+Fix "'License :: OSI Approved :: ISC License' is not a valid classifier" error running "python3 -m twine upload dist/*" [FA-84].
+
+
+## 0.1.11 (2024-10-17)
 ---
 
 ### New